pyproject.toml
[project]
name = "flask-shop-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-mail>=0.10.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "werkzeug>=3.1.4",
    "wtforms>=3.2.1",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "flake8>=7.3.0",
    "pytest>=9.0.2",
    "pytest-flask>=1.3.0",
]
